+++ src/lib.rs
@@ -1,3 +1,5 @@
+mod attention;
 mod softmax;
 
+pub use attention::flash_attention;
 pub use softmax::online_softmax;